server/requirements.txt
openenv-core>=0.2.2
uvicorn>=0.24.0
fastapi>=0.115.0